Tidy up os_environ fixture
In response to [^1] and [^2].

[^1]: https://github.com/alphagov/notifications-template-preview/pull/651#discussion_r856143150
[^2]: https://github.com/alphagov/notifications-template-preview/pull/651#discussion_r856149694
